Fix cursor issues
